Make cluster opt.ReadOnly be a *bool
This allows us to set it in the client to false if we truly want it to be false.
Not setting it means it can be set to true by setting RouteRandomly or RouteByLatency as usual.
This is to support Redislabs which don't allow the READONLY command at all (should be a noop, but alas)